Changed html to look like prototype
So the menubar is from tailwind, and will need js. We will implement towards an app, and evaluate that tmmr. The image and the gradient usage with tailwind for me was scuffed, The gradiant is horizontal, and making it linear for me requried a 90 deg rotation that would then screw the length and width, and make it nearly impossible to scale right with the image. The new issue is that the square class I made in order to go over the image as a gradient, does not use the length of the image in order to specificy the end point of the gradient line. Might be related to how I am setting up the content box. Besides that, nothing else.